Adds a c_str method to StringRef
Current class does not ensure the string is null terminated. As there is
no defined error to actually check for a null terminator, we determine
if it is null terminated based on the constructor type. Then if on c_str
it is not known to be null terminated, we use a passed in buffer to copy
the string and ensure null termination